Found a nicer way to use data from route params.
It turns out there's an ownProps object if you want to inherit other
props that are passed down outside of Redux. Since react-router uses
that to propogate props, I modifiy my map in the container to grab
that value and stuff it into subreddit. This works b/c all the redux
state uses the value for is to collect the name from the form.

This is probably about as good as I can do without using react-router-redux.